% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CERN.R
\docType{data}
\name{CERN}
\alias{CERN}
\title{277 measurements of the cross sections for
\eqn{\pi^{-}p} collision (nuclear
physics).}
\format{A numeric vector with 277 elements.}
\source{
\url{https://link.springer.com/article/10.1007/BF02683433}
}
\usage{
CERN
}
\description{
277 measurements of the cross sections for
\eqn{\pi^{-}p} collision (nuclear
physics).
}
\keyword{datasets}
