% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_demo.R
\name{forecast_demo}
\alias{forecast_demo}
\title{Forecasting demo using cics_unif_explicit_smooth.}
\usage{
forecast_demo()
}
\value{
a forecast result
}
\description{
Forecasting demo using cics_unif_explicit_smooth.
}
\examples{

forecast_demo()

}
