% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICS_smoothing.R
\name{hermite_bf_matrix}
\alias{hermite_bf_matrix}
\title{Construct 4 Hermite basis functions.}
\usage{
hermite_bf_matrix(u, v)
}
\arguments{
\item{u}{a left border of interval [\code{u},\code{v}].}

\item{v}{a right border of interval [\code{u},\code{v}], \code{u}\eqn{\le}\code{v}.}
}
\value{
The matrix of 4 Hermite basis functions' coefficients.
}
\description{
\code{hermite_bf_matrix} constructs matrix of Hermite basis functions' coefficients on
[\code{u},\code{v}], that is the matrix of 4 cubic polynomials' coefficients of
one-component Hermite cubic spline.
}
\examples{

hermite_bf_matrix(0,1)
hermite_bf_matrix(-2,3)
}
