% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICS_smoothing.R
\name{tridiag_inv_general}
\alias{tridiag_inv_general}
\title{Construct inverse of a general tridiagonal matrix.}
\usage{
tridiag_inv_general(T, n)
}
\arguments{
\item{T}{a tridiagonal matrix.}

\item{n}{an order of given tridiagonal matrix.}
}
\value{
The inverse of matrix T.
}
\description{
\code{tridiag_inv_general} constructs inverse of a general tridiagonal matrix \code{T} of order \code{n},
using Usmani's theorem.
}
\examples{

tridiag_inv_general(matrix(c(1,3,5,-2,0,8,7,6,6),3,3),3)
tridiag_inv_general(matrix(c(1,4,0,-9),2,2),2)
}
