% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICS_smoothing.R
\name{tridiag_inv_unif_by_sums}
\alias{tridiag_inv_unif_by_sums}
\title{Construct inverse of a tridiagonal matrix T_n(a,b,a).}
\usage{
tridiag_inv_unif_by_sums(n, a, b)
}
\arguments{
\item{n}{an order of given tridiagonal matrix.}

\item{a}{a value of tridiagonal matrix elements that are off-diagonal.}

\item{b}{a value of tridiagonal matrix diagonal elements.}
}
\value{
The inverse of matrix \code{T}_{\code{n}}(\code{a},\code{b},\code{a}).
}
\description{
\code{tridiag_inv_unif_by_sums} constructs inverse of a regular tridiagonal matrix \code{T}_{\code{n}}(\code{a},\code{b},\code{a})
with constant entries by a special algorithm using sums of matrix elements.
}
\examples{
tridiag_inv_unif_by_sums(5, 1, 4)
tridiag_inv_unif_by_sums(9, 10, -1)
}
