% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICS_smoothing.R
\name{cics_explicit}
\alias{cics_explicit}
\title{Construct the explicit form of non-uniform clamped interpolating cubic spline (NcICS).}
\usage{
cics_explicit(
  uu,
  yy,
  d,
  clrs = c("blue", "red"),
  xlab = NULL,
  ylab = NULL,
  title = "Spline"
)
}
\arguments{
\item{uu}{a vector of arbitrary knots (ordered ascendingly), with magnitude \code{n+2}, \code{n}\eqn{\ge}\code{1}.}

\item{yy}{a vector of function values pertaining to knots in \code{uu}.}

\item{d}{a vector of two values of derivative, in the first and the last knot of \code{uu}.}

\item{clrs}{a vector of colours that are used alternately to plot the graph of spline's components.}

\item{xlab}{a title (optional parameter) for the \code{x} axis.}

\item{ylab}{a title (optional parameter) for the \code{y} axis.}

\item{title}{a title (optional parameter) for the plot.}
}
\value{
a list with components
\item{spline}{matrix, whose \code{i}-th row contains coefficients of non-uniform ICS's \code{i}-th component.}
\item{spline_toPolynomial}{list of NcICS's components string representations.}
\item{B}{\code{4}-element array of \code{(n+1)x(n+4)} matrices, whereas element in \code{i}-th row
and \code{j}-th column of \code{l}-th matrix contains coefficient by \code{x^{l-1}} of cubic polynomial that is in \code{i}-th row
and \code{j}-th column of matrix \code{B} from spline's explicit form \deqn{S=B.\gamma.}}
\item{gama}{\eqn{\gamma=} vector of spline coefficients - function values and exterior-knot derivatives that
 takes part in the explicit form \eqn{S=B.\gamma}.}
\item{aux_BF}{A basis function of the spline}
\item{aux_tridiag_inverse}{An inverse of the tridiagonal matrix used for spline derivatives construction}
}
\description{
\code{cics_explicit} constructs the explicit form of non-uniform clamped interpolating cubic spline
(via Hermite cubic spline) for knots \code{uu}, function values \code{yy} and exterior-knot derivatives
\code{d}.
}
\examples{

cics_explicit(c(0,2,7),c(4,5,-1),c(0,-2),c('blue','red'))
cics_explicit(c(-2,-1,3,10,13),c(-5,-4,2,5,2),c(1,-3),c('green','yellow'))
cics_explicit(CERN[,1],CERN[,2], d=c(0,-2), xlab="X axis", ylab="Y axis")
}
