% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICS_smoothing.R
\name{cics_unif_explicit}
\alias{cics_unif_explicit}
\title{Construct the explicit form of uniform clamped interpolating cubic spline (UcICS).}
\usage{
cics_unif_explicit(
  uumin,
  uumax,
  yy,
  d,
  clrs = c("blue", "red"),
  xlab = NULL,
  ylab = NULL,
  title = "Spline"
)
}
\arguments{
\item{uumin}{a starting knot.}

\item{uumax}{an ending knot.}

\item{yy}{a vector of function values pertaining to knots in \code{uu}.}

\item{d}{a vector of two values of derivative, in the first and the last knot of \code{uu}.}

\item{clrs}{a vector (optional parameter) of colours that are used alternately to plot the graph of spline's components.}

\item{xlab}{a title (optional parameter) for the \code{x} axis.}

\item{ylab}{a title (optional parameter) for the \code{y} axis.}

\item{title}{a title (optional parameter) for the plot.}
}
\value{
a list of spline components
\item{spline}{matrix, whose \code{i}-th row contains coefficients of uniform ICS's \code{i}-th component.}
\item{spline_toPolynomial}{list of UcICS's components string representations.}
\item{B}{\code{4}-element array of \code{(n+1)x(n+4)} matrices, whereas element in \code{i}-th row
and \code{j}-th column of \code{l}-th matrix contains coefficient by \code{x^{l-1}} of cubic polynomial that is in \code{i}-th row
and \code{j}-th column of matrix \code{B} from spline's explicit form \deqn{S=B.\gamma.}}
\item{gama}{\eqn{\gamma=} vector of spline coefficients - function values and exterior-knot derivatives that
 takes part in the explicit form \eqn{S=B.\gamma}.}
}
\description{
\code{cics_unif_explicit} constructs the explicit form of uniform clamped interpolating cubic spline
(via Hermite cubic spline) for knots \code{uu}, function values \code{yy} and exterior-knot derivatives
\code{d}.
}
\examples{

cics_unif_explicit(
head(CERN$x, n=1),
tail(CERN$x, n=1),
CERN$y,
c(0,0),
xlab="X axis",
ylab="Y axis"
)

}
